gpu_unai: Factored out gpu_unai macros/defines into new gpu_unai.h header
New header contains lots of code common to original gpu_unai standalone
plugin that was duplicated in gpulib_if.cpp version.
